Specified that rDock was developed at RiboTargets
That occurence of rDock was wrongly generalized to |Dock| (so it can be
rDock or RxDock depending on the value of the replacement text).

Thanks to Sergio Ruiz Carmona <Sergio.RuizCarmona@baker.edu.au> for
spotting this mistake and reporting it.

Also changed one wrong occurence of rDock in the User guide to |Dock|,
where we link to download links for RxDock.